Refactored authorization with before filter.

+++ app/controllers/questions_controller.rb
@@ -1,19 +1,20 @@
 class QuestionsController < ApplicationController
+  before_filter :authorize, except: [:index, :show]
+
   def index
     @questions = Question.order('created_at DESC')
   end
 
   def new
+    # authorize
+
     @question = Question.new
   end
 
   def create
     params[:question][:author] = session[:user_id]
     @question = Question.new(question_params)
-    if !current_user
-      flash[:notice] = "Please sign in to ask a question."
-      render :new
-    elsif @question.save
+   if @question.save
       redirect_to "/questions/#{@question.id}"
     else
       flash.now[:notice] = "Could not save your question."
@@ -27,11 +28,11 @@ class QuestionsController < ApplicationController
   end
 
   def edit
+    # authorize
+
     @question = Question.find(params[:id])
-    if !current_user
-      flash[:notice] = "Please sign in to edit questions."
-      redirect_to "/questions/#{@question.id}"
-    elsif current_user.id != @question.author.to_i
+
+    if current_user.id != @question.author.to_i
       flash[:notice] = "You are not allowed to edit this question."
       redirect_to "/questions/#{@question.id}"
     end
@@ -48,12 +49,21 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
+    authorize
+
+
     question = Question.find(params[:id])
-    question.answers.each do |answer|
-      answer.destroy
+
+    if current_user.id == @question.author.to_i
+      question.answers.each do |answer|
+        answer.destroy
+      end
+      question.destroy
+      redirect_to action: :index
+    else
+      flash[:notice] = "You are not allowed to delete this question."
+      redirect_to "/questions/#{@question.id}"
     end
-    question.destroy
-    redirect_to action: :index
   end
 
   private
